Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/xhr-in-moduleworker-expected.txt


FAIL Base URL in module dedicated workers: XHR assert_equals: expected "gamma\n" but got "beta\n"
